+++ game-of-life.py
@@ -1,6 +1,6 @@
 #!venv/bin/python
 
-from src.root import main
+from src.app import main
 
 if __name__ == "__main__":
     main()
